Fixed bug where signal manager buffer did not flush due to windows file lock. SignalManager now writes partial parquet files (up to 8 MB/user defined). When close() is called (such as when a runtime manager context closes) the partials are read back in and streamed together.

+++ src/mujoco_mojo/runtime/video_recorder.py
@@ -117,7 +117,7 @@ class VideoRecorder:
     show_traces: bool = False
     """Whether to render `Tracer` trails (passed via `custom_traces` in `capture_frame`)."""
 
-    fps: int = 30
+    fps: float = 30
     """Target frame rate of the output video. Frames are sampled every `1/fps` seconds of simulation time."""
 
     playback_speed: float = 1.0
